+++ frontend/src/app/main/ui/workspace/sidebar/options/shapes/image.cljs
@@ -8,9 +8,11 @@
   (:require
    [app.main.ui.workspace.sidebar.options.menus.blur :refer [blur-menu]]
    [app.main.ui.workspace.sidebar.options.menus.constraints :refer [constraint-attrs constraints-menu]]
+   [app.main.ui.workspace.sidebar.options.menus.fill :refer [fill-attrs fill-menu]]
    [app.main.ui.workspace.sidebar.options.menus.layer :refer [layer-attrs layer-menu]]
    [app.main.ui.workspace.sidebar.options.menus.measures :refer [measure-attrs measures-menu]]
    [app.main.ui.workspace.sidebar.options.menus.shadow :refer [shadow-menu]]
+   [app.main.ui.workspace.sidebar.options.menus.stroke :refer [stroke-attrs stroke-menu]]
    [rumext.alpha :as mf]))
 
 (mf/defc options
@@ -19,7 +21,9 @@
         type (:type shape)
         measure-values (select-keys shape measure-attrs)
         layer-values (select-keys shape layer-attrs)
-        constraint-values (select-keys shape constraint-attrs)]
+        constraint-values (select-keys shape constraint-attrs)
+        fill-values (select-keys shape fill-attrs)
+        stroke-values (select-keys shape stroke-attrs)]
     [:*
      [:& measures-menu {:ids ids
                         :type type
@@ -31,6 +35,14 @@
      [:& layer-menu {:ids ids
                      :type type
                      :values layer-values}]
+
+     [:& fill-menu {:ids ids
+                    :type type
+                    :values fill-values}]
+
+     [:& stroke-menu {:ids ids
+                      :type type
+                      :values stroke-values}]
 
      [:& shadow-menu {:ids ids
                       :values (select-keys shape [:shadow])}]